got auth middleware working
