+++ backend/middleware/authMiddleware.js
@@ -3,7 +3,7 @@ const serviceAccount = require("../configs/serivceAccountKey");
 const admin = require("firebase-admin");
 
 const SECRET_KEY = process.env.JWT_SECRET_KEY;
-const JWT_EXPIRY = 300;
+const JWT_EXPIRY = 60 * 60 * 24 * 7;
 
 admin.initializeApp({
     credential: admin.credential.cert(serviceAccount)
@@ -17,23 +17,23 @@ const createToken = (id) => {
 
 const checkForFirebaseToken = async (req, res, next) => {
     console.log("checking for token");
-    //const firebaseToken = req.headers.authorization?.split("Bearer ")[1];
+    const firebaseToken = req.headers.authorization?.split("Bearer ")[1];
 
-    // if (!firebaseToken) {
-    //     console.log("didn't find token")
-    //     return res.status(401).json({ error: "No Firebase token provided" });
-    // }
+    if (!firebaseToken) {
+        console.log("didn't find token")
+        return res.status(401).json({ error: "No Firebase token provided" });
+    }
 
     try {
         console.log("found token, attempting decoding")
-        //const decodedFirebaseToken = await admin.auth().verifyIdToken(firebaseToken);
+        const decodedFirebaseToken = await admin.auth().verifyIdToken(firebaseToken);
         console.log("Attempting to create custom token")
-        // const customToken = createToken(decodedFirebaseToken.uid);
-        // console.log(customToken);
+        const customToken = createToken(decodedFirebaseToken.uid);
+        console.log(customToken);
         res.status(200)
-           .cookie("customToken", JSON.stringify({ id: "sam" }), {
+           .cookie("customToken", customToken, {
                 httpOnly: true,
-                secure: false,
+                secure: true,
                 sameSite: "lax",
                 maxAge: JWT_EXPIRY * 1000 
             })
@@ -47,21 +47,16 @@ const checkForFirebaseToken = async (req, res, next) => {
     }
 }
 
-const checkForCustomToken = (req, res, next) => {
+const checkForCustomToken = async (req, res, next) => {
     const token = req.cookies.customToken;
     console.log(req.cookies);
     if (!token) {
         return res.status(401).json({ error: "Authentication required" });
     }
     try {
-        jwt.verify(token, SECRET_KEY, (err) => {
-            if (err) {
-                console.log(err.message);
-                return res.status(401).json({ error: "Couldn't verify token" });
-            } else {
-                next();
-            }
-        })
+        const decoded = await jwt.verify(token, SECRET_KEY);
+        req.user = decoded;
+        next();
     } catch (err) {
         console.log(err.message);
         return res.status(401).json({ error: "Invalid or expired token" });
